requirements.txt
ansible==5.6.0
ansible-core==2.12.4
ansible-runner==2.1.3
anyio==3.5.0
asgiref==3.5.1
bcrypt==3.2.0
boto3==1.22.5
botocore==1.25.5
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
cryptography==36.0.2
dnspython==2.2.1
docutils==0.18.1
email-validator==1.2.1
fastapi==0.75.2
h11==0.13.0
httptools==0.4.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
jmespath==1.0.0
lockfile==0.12.2
MarkupSafe==2.1.1
orjson==3.6.8
packaging==21.3
paramiko==2.10.3
pexpect==4.8.0
ptyprocess==0.7.0
pycparser==2.21
pydantic==1.9.0
Pygments==2.13.0
PyNaCl==1.5.0
pyparsing==3.0.8
python-daemon==2.3.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.27.1
resolvelib==0.5.4
rich==12.6.0
s3transfer==0.5.2
shellingham==1.5.0
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typer==0.7.0
typing_extensions==4.2.0
ujson==5.2.0
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
